refactoring of results and events types definitions
