+++ include/libaktualizr/campaign.h
