+++ include/libaktualizr/config.h
@@ -9,8 +9,8 @@
 #include <boost/filesystem.hpp>
 #include <boost/program_options.hpp>
 #include <boost/property_tree/ini_parser.hpp>
+
 #include "libaktualizr/types.h"
-#include "utilities/utils.h"
 
 // Try to keep the order of config options the same as in Config::writeToStream()
 // and Config::updateFromPropertyTree() in config.cc.
