+++ src/libaktualizr/campaign/campaign.cc
@@ -1,4 +1,4 @@
-#include "campaign/campaign.h"
+#include "libaktualizr/campaign.h"
 
 namespace campaign {
