+++ src/libaktualizr/primary/CMakeLists.txt
@@ -8,10 +8,8 @@ set(SOURCES aktualizr.cc
 set(HEADERS secondary_config.h
             aktualizr.h
             aktualizr_helpers.h
-            events.h
             initializer.h
             reportqueue.h
-            results.h
             secondaryinterface.h
             secondary_provider.h
             sotauptaneclient.h)
