+++ src/libaktualizr/primary/aktualizr.cc
@@ -4,7 +4,7 @@
 
 #include <sodium.h>
 
-#include "primary/events.h"
+#include "libaktualizr/events.h"
 #include "utilities/timer.h"
 
 using std::make_shared;
